add features to test-utils
