+++ frame/mra-tls/Cargo.toml
@@ -26,7 +26,7 @@ yasna = { rev = "sgx_1.1.3", git = "https://github.com/mesalock-linux/yasna.rs-s
 bit-vec = { version = "0.6.1", default-features = false }
 num-bigint = { version = "0.2", git = "https://github.com/mesalock-linux/num-bigint-sgx" }
 chrono = { version = "0.4", git = "https://github.com/mesalock-linux/chrono-sgx" }
-test-utils = { path = "../../tests/utils" }
+test-utils = { path = "../../tests/utils", default-features = false, features = ["sgx"] }
 pem = { version = "0.8.2", git = "https://github.com/mesalock-linux/pem-rs-sgx" }
 tracing = { version = "0.1", default-features = false }
 log = { git = "https://github.com/mesalock-linux/log-sgx" }
