+++ modules/anonify-enclave/Cargo.toml
@@ -15,7 +15,7 @@ frame-common = { path = "../../frame/common", default-features = false, features
 frame-mra-tls = { path = "../../frame/mra-tls" }
 remote-attestation = { path = "../../frame/remote-attestation" }
 anonify-ecall-types = { path = "../anonify-ecall-types", default-features = false, features = ["sgx"] }
-test-utils = { path = "../../tests/utils" }
+test-utils = { path = "../../tests/utils", default-features = false, features = ["sgx"] }
 serde_json = { rev = "sgx_1.1.3", git = "https://github.com/mesalock-linux/serde-json-sgx" }
 serde = { git = "https://github.com/mesalock-linux/serde-sgx.git" } # Don't specify version due to serde_json dependency
 bincode = { rev = "sgx_1.1.3", git = "https://github.com/mesalock-linux/bincode-sgx" }
