+++ tests/units/enclave/Cargo.toml
@@ -15,7 +15,7 @@ frame-config = { path = "../../../frame/config", default-features = false, featu
 anonify-enclave = { path = "../../../modules/anonify-enclave", default-features = false }
 sgx_tstd = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git"}
 lazy_static = { version = "1.4", features = ["spin_no_std"] }
-test-utils = { path = "../../utils" }
+test-utils = { path = "../../utils", default-features = false, features = ["sgx"] }
 
 [features]
 default = ["backup-enable", "runtime_enabled"]
