+++ tests/utils/Cargo.toml
@@ -5,7 +5,17 @@ authors = ["LayerX Labs <div-labs@layerx.co.jp>"]
 edition = "2018"
 
 [dependencies]
-inventory = { rev = "sgx_1.1.3", git = "https://github.com/mesalock-linux/inventory-sgx"}
-serde = { rev = "sgx_1.1.3", git = "https://github.com/mesalock-linux/serde-sgx.git", features = ["derive"] }
-sgx_tstd = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git", features = ["net"] }
-test-utils-proc-macro = { path = "./proc-macro" }
+inventory = { rev = "sgx_1.1.3", git = "https://github.com/mesalock-linux/inventory-sgx", optional = true }
+serde = { rev = "sgx_1.1.3", git = "https://github.com/mesalock-linux/serde-sgx.git", features = ["derive"], optional = true }
+sgx_tstd = { rev = "v1.1.3", git = "https://github.com/apache/teaclave-sgx-sdk.git", features = ["net"], optional = true }
+test-utils-proc-macro = { path = "./proc-macro", optional = true  }
+
+[features]
+default = ["std"]
+std = []
+sgx = [
+    "inventory",
+    "serde",
+    "sgx_tstd",
+    "test-utils-proc-macro",
+]
